requirements.txt
# Telegram bilan ishlash
python-telegram-bot==20.3

# Sozlamalar uchun
python-dotenv==1.0.0

# Ma'lumotlar bilan ishlash
pandas==2.0.3
openpyxl==3.1.2
numpy==1.24.3

# Qo'shimcha funksional
requests==2.31.0
apscheduler==3.10.1
python-dateutil==2.8.2

# Loglash
loguru==0.7.0

# Deploy uchun
gunicorn==20.1.0
flask==2.2.3

# PostgreSQL uchun (agar ishlatsangiz)
psycopg2-binary==2.9.5

pyTelegramBotAPI==4.15.3
pytz==2024.1

pymongo==4.5.0
certifi==2023.7.22